requirements.txt
algoliasearch==1.13.0
algoliasearch-django==1.4.1
appnope==0.1.0
asn1crypto==0.22.0
astroid==1.5.3
bcrypt==3.1.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==2.0
decorator==4.1.1
defusedxml==0.5.0
Django==1.11.3
django-appconf==1.0.2
django-braces==1.11.0
django-cachalot==1.5.0
django-compressor==2.1.1
django-debug-toolbar==1.8
django-extensions==1.8.1
django-flexible-images==1.0.7
django-historylinks==1.1.0
django-jinja==2.2.1
django-reversion==2.0.9
django-suit==0.2.25
django-watson==1.3.1
django-webpack-loader==0.5.0
docopt==0.6.2
ecdsa==0.13
Fabric3==1.13.1.post1
fabric3-virtualenv==0.3.1
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.8
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
onespacemedia-cms==3.0.0
onespacemedia-cms-redirects==0.0.2
onespacemedia-server-management==3.0.1
opbeat==3.5.2
paramiko==1.16.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
pyasn1==0.2.3
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.2
pylint==1.7.2
pylint-django==0.7.2
pylint-mccabe==0.1.3
pylint-plugin-utils==0.2.6
PyNaCl==1.1.2
pyOpenSSL==17.2.0
python-memcached==1.58
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
requests==2.18.1
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sorl-thumbnail==12.3
sqlparse==0.2.3
tinypng==3.0.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.10
